+++ PowerSync/PowerSync.Common/Client/Sync/Bucket/BucketStorageAdapter.cs
@@ -18,6 +18,19 @@ public static class PowerSyncControlCommand
     public const string NOTIFY_TOKEN_REFRESHED = "refreshed_token";
     public const string NOTIFY_CRUD_UPLOAD_COMPLETED = "completed_upload";
     public const string UPDATE_SUBSCRIPTIONS = "update_subscriptions";
+
+    public const string TARGET_CHECKPOINT_REQUEST_ID = "target_checkpoint_request_id";
+
+    /// <summary>
+    /// An `established` or `end` event for response streams.
+    /// </summary>
+    public const string CONNECTION_STATE = "connection";
+}
+
+public static class PowerSyncControlConnectionState
+{
+    public const string ESTABLISHED = "established";
+    public const string END = "end";
 }
 
 public class Checkpoint
@@ -117,16 +130,14 @@ public class BucketStorageEvents : EventManager
 public interface IBucketStorageAdapter : ICloseable
 {
     BucketStorageEvents Events { get; }
-    Task Init();
 
     Task<CrudEntry?> NextCrudItem();
     Task<bool> HasCrud();
     Task<CrudBatch?> GetCrudBatch(int limit = 100);
 
-    Task<bool> HasCompletedSync();
     Task<bool> UpdateLocalTarget(Func<Task<string>> callback);
 
-    string GetMaxOpId();
+    Task HandleCrudCheckpoint(long lastClientId, string? writeCheckpoint = null);
 
     /// <summary>
     /// Get a unique client ID.